Move the hal resource and link classes to be outside of hal
